Uploaded 1013A http://www.indiedb.com/downloads/start/88391?referer=http%3A%2F%2Fwww.indiedb.com%2Fmods%2Fdeveks-mod%2Fdownloads updated combat AI main targetting choices. escort ships should protect their escorts more. working on fleets and global military AI. this should help get the AI good invasion fleets. The AI can use troopbays instead of troops if it has the designs. it will count each bay as a single troop. Disabled some test code that would cause a crash in mods on new games. added some locking around troop counting. added a try catch instead of locking around some of the freighter code as i could not get it to lock and a null value is ok in that spot. removed some more conversions to doubles. Have to to remember to add the f on 0.0. thats why the decomplier upgraded everything to doubles...
--HG--
branch : develop

+++ Ship_Game/UniverseScreen.cs
@@ -2187,7 +2187,7 @@ namespace Ship_Game
                 {
                     system.DangerTimer -= realTime;
                     system.DangerUpdater -= realTime;
-                    if ((double)system.DangerUpdater < 0.0)
+                    if (system.DangerUpdater < 0.0)
                     {
                         system.DangerUpdater = 10f;
                         system.DangerTimer = (double)this.player.GetGSAI().ThreatMatrix.PingRadarStr(system.Position, 100000f * UniverseScreen.GameScaleStatic, this.player) <= 0.0 ? 0.0f : 120f;
@@ -2195,7 +2195,7 @@ namespace Ship_Game
                     system.combatTimer -= realTime;
 
 
-                    if ((double)system.combatTimer <= 0.0)
+                    if (system.combatTimer <= 0.0)
                         system.CombatInSystem = false;
                     bool viewing = false;
                     this.ScreenManager.GraphicsDevice.Viewport.Project(new Vector3(system.Position, 0.0f), this.projection, this.view, Matrix.Identity);
